Attach server to scans.
The data that comes back with scans has some info about the server,
including its hostname. But there is additional data about the server
that we may want, including the EC2 instance-id (if it exists). We
already have the server data, so this adds zero requests. It just uses
slightly more RAM now because that server data is attached to each scan.